refactor: consolidate duplicate modules
