--- data/badges.js
@@ -1,33 +1,0 @@
-export const BADGE_LIST = [
-  {
-    id: 'premiumMember',
-    icon: 'diamond-outline',
-    title: 'Premium Member',
-    desc: 'Subscribed to Pinged Premium.',
-    premium: true,
-  },
-  {
-    id: 'firstWin',
-    icon: 'trophy-outline',
-    title: 'First Win',
-    desc: 'Win your first game.',
-  },
-  {
-    id: 'perfectGame',
-    icon: 'star-outline',
-    title: 'Perfect Game',
-    desc: 'Win a match without mistakes.',
-  },
-  {
-    id: 'dailyStreak',
-    icon: 'flame-outline',
-    title: 'Daily Streak',
-    desc: 'Play 7 days in a row.',
-  },
-  {
-    id: 'socialButterfly',
-    icon: 'people-outline',
-    title: 'Social Butterfly',
-    desc: 'Join your first community event.',
-  },
-];
