--- data/games.js
@@ -1,1 +1,0 @@
-export { allGames, gameMap } from '../constants/games';
